+++ ReleaseNotes.md
@@ -6,6 +6,7 @@
 
 * TGR-1110: Zion - XML-structures with nodes named 'text' can now be correctly serialized
 * TGR-971: Tiger Proxy UI - when applying filters, the list of senders and receivers is limited to the hosts which send or receive more than two messages. An additional checkbox allows displaying the full list of senders and receivers.
+* TGR-1108: Allow nested JexlExpression in RbelPaths (up to one level!)
 * TGR-1112: Fixed that log contained a lot of "curl command unable to parse log" lines
 
 -------